+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CDTDatastore CHANGELOG
 
+## Unreleased
+- [FIXED] Bug which prevented CDTDatastore instances from being closed after calling index/query
+  methods.
+
 ## 2.0.2 (2018-04-30)
 - [FIXED] Bug which prevented `find` queries from executing
   successfully against indexes with certain names.